NFDIV-4744: Fix issue with validation errors
