+++ src/main/steps/applicant1/interim-applications/search-gov-records/partner-address-additional/template.njk
@@ -5,7 +5,7 @@
 {% extends "common/page.njk" %}
 
 {% block page_content %}
-  <h1 class="govuk-heading-l" id="whichGovDepartments">{{ title }}</h1>
+  <h1 class="govuk-heading-l" id="partnerAddressAdditional">{{ title }}</h1>
   <h2 class="govuk-heading-m">{{ address }}</h2>
 {% endblock %}
